+++ miracle/scripts/tests/test_analysis.py
@@ -7,7 +7,9 @@ from miracle.models import (
 )
 from miracle.scripts import analysis
 
-TEST_START = datetime.utcfromtimestamp(1469400000)
+NOW = datetime.utcnow()
+USER_CREATED = NOW - timedelta(days=20)
+TEST_START = NOW - timedelta(days=14)
 
 
 def test_weekly_recurrence(db):
@@ -16,8 +18,11 @@ def test_weekly_recurrence(db):
         url2 = URL(**URL.from_url('https://localhost/'))
         url3 = URL(**URL.from_url('https://example.com/'))
         url4 = URL(**URL.from_url('https://example.com/other'))
-        user1 = User(token='user1')
-        user2 = User(token='user2')
+        user1 = User(token='user1', created=USER_CREATED)
+        user2 = User(token='user2', created=USER_CREATED)
+        user3 = User(token='user3', created=NOW)
+        # Fourth user without any sessions
+        session.add(User(token='user4', created=USER_CREATED))
         session.add_all([
             # URL visited three times in a week
             Session(user=user1, url=url1, start_time=TEST_START),
@@ -48,8 +53,15 @@ def test_weekly_recurrence(db):
             Session(user=user1, url=url4,
                     start_time=TEST_START + timedelta(days=12)),
             # Second user with too few sessions
+            Session(user=user2, url=url3, start_time=TEST_START),
             Session(user=user2, url=url3,
-                    start_time=TEST_START + timedelta(days=2)),
+                    start_time=TEST_START + timedelta(days=8)),
+            # Third user with too little participation time
+            Session(user=user3, url=url3, start_time=TEST_START),
+            Session(user=user3, url=url3,
+                    start_time=TEST_START + timedelta(days=1)),
+            Session(user=user3, url=url3,
+                    start_time=TEST_START + timedelta(days=10)),
         ])
         session.flush()
         assert analysis.weekly_recurrence(db) == [(0, 1), (2, 1)]